add file parser.rb
